# internal/pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Library server",
        "title": "Library API",
        "contact": {
            "name": "Kirill Efremenko",
            "email": "kirikozavrr@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/library",
    "paths": {
        "/authors/add": {
            "post": {
                "description": "Adding an author",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authors"
                ],
                "summary": "Add an author",
                "parameters": [
                    {
                        "description": "Author information",
                        "name": "author",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.AddAuthorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.AddAuthorResponse"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/authors/all": {
            "get": {
                "description": "Getting all authors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authors"
                ],
                "summary": "Get All authors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetAllAuthorsResponse"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/authors/top": {
            "get": {
                "description": "Getting a top of author with your limit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authors"
                ],
                "summary": "Get top authors",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "This is how many authors need to get",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetTopAuthorsResponse"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/books/add": {
            "post": {
                "description": "Adding a book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Add a book",
                "parameters": [
                    {
                        "description": "Book information with author id",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.AddBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.AddBookResponse"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/books/all": {
            "get": {
                "description": "Getting all Books",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Get All Books",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetAllBooksResponse"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.AddAuthorRequest": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                }
            }
        },
        "controller.AddAuthorResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.DataAuthor"
                },
                "error": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "controller.AddBookRequest": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "controller.AddBookResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.DataBook"
                },
                "error": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "controller.DataAuthor": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.AuthorDTO"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.DataAuthors": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AuthorDTO"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.DataBook": {
            "type": "object",
            "properties": {
                "book": {
                    "$ref": "#/definitions/models.BookDTO"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.DataBooks": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BookDTO"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controller.GetAllAuthorsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.DataAuthors"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "controller.GetAllBooksResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.DataBooks"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "controller.GetTopAuthorsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.DataAuthors"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.AuthorDTO": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BookDTO"
                    }
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "models.BookDTO": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.AuthorDTO"
                },
                "id": {
                    "type": "integer"
                },
                "rented": {
                    "$ref": "#/definitions/models.UserDTO"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.UserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "rented_books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BookDTO"
                    }
                },
                "user_status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}